refactored general build/testing utils into their own directory and further refactored the jsxPragma situation
