+++ package.json
@@ -32,10 +32,10 @@
   "scripts": {
     "build": "npm run prod && npm run dev",
     "prod": "webpack -p",
-    "dev": "webpack -d --devtool hidden --output-file=tota11y.js",
-    "lint": "eslint index.js plugins test",
+    "dev": "webpack --config utils/webpack.config.babel.js -d --devtool hidden --output-file=tota11y.js",
+    "lint": "eslint index.js plugins test utils",
     "test": "mocha --require test/babel-hook test/*.js",
-    "live-test": "webpack-dev-server --hot --inline"
+    "live-test": "webpack-dev-server --config utils/webpack.config.babel.js --hot --inline"
   },
   "license": "MIT"
 }
