+++ test/babel-hook.js
@@ -4,8 +4,12 @@
  * This sets up babel with options, and stores our JSX transpile target.
  */
 
+var options = require("../utils/options");
+
 // Register all future "require"s with babel
-require("babel/register");
+require("babel/register")({
+    jsxPragma: options.jsxPragma,
+});
 
 // Store our custom JSX transpile target as a global
-E = require("../element.js");       // eslint-disable-line no-undef
+global[options.jsxPragma] = require("../utils/element.js");
